Allow multiple header values for the same key
The Email struct can have multiple values added for the same name. This
commit updates the `put_header` function on the Email module by allowing
header values to be added as a list and combining multiple values for
the same name into lists. The Mailer module also has a function to
normalize those headers into a list of two element tuples to can be sent
to the http client library based on the strategy elected by the caller.

Resolves #469, #343, #360